+++ src/components/DashboardsBar/Content.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types'
 import React, { useState } from 'react'
 import { connect } from 'react-redux'
 import { Redirect, withRouter } from 'react-router-dom'
-import { useOnlineStatus } from '../../modules/useFakeOnlineStatus.js'
+import { useOnlineStatus } from '../../online-status-fakes/useFakeOnlineStatus.js'
 import { sGetAllDashboards } from '../../reducers/dashboards.js'
 import { sGetDashboardsFilter } from '../../reducers/dashboardsFilter.js'
 import { sGetSelectedId } from '../../reducers/selected.js'
