+++ src/components/DropdownButton/DropdownButton.js
@@ -1,7 +1,7 @@
 import { Button, Layer, Popper } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React, { useRef } from 'react'
-import { OfflineTooltip } from '../FakeOfflineTooltip.js'
+import { OfflineTooltip } from '../../online-status-fakes/FakeOfflineTooltip.js'
 import { ArrowDown, ArrowUp } from './assets/Arrow.js'
 import styles from './DropdownButton.module.css'
