+++ src/components/Item/VisualizationItem/Visualization/MapPlugin.js
@@ -2,7 +2,7 @@ import i18n from '@dhis2/d2-i18n'
 import PropTypes from 'prop-types'
 import React, { useState, useEffect } from 'react'
 import { MAP } from '../../../../modules/itemTypes.js'
-import { useOnlineStatus } from '../../../../modules/useFakeOnlineStatus.js'
+import { useOnlineStatus } from '../../../../online-status-fakes/useFakeOnlineStatus.js'
 import getVisualizationContainerDomId from '../getVisualizationContainerDomId.js'
 import { isElementFullscreen } from '../isElementFullscreen.js'
 import DefaultPlugin from './DefaultPlugin.js'
