+++ src/pages/edit/LayoutModal.js
@@ -12,9 +12,9 @@ import {
 import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React, { useEffect, useState } from 'react'
-import { OfflineTooltip } from '../../components/FakeOfflineTooltip.js'
 import { GRID_COLUMNS } from '../../modules/gridUtil.js'
-import { useOnlineStatus } from '../../modules/useFakeOnlineStatus.js'
+import { OfflineTooltip } from '../../online-status-fakes/FakeOfflineTooltip.js'
+import { useOnlineStatus } from '../../online-status-fakes/useFakeOnlineStatus.js'
 import { LayoutFixedIcon } from './assets/LayoutFixed.js'
 import { LayoutFreeflowIcon } from './assets/LayoutFreeflow.js'
 import classes from './styles/LayoutModal.module.css'
