+++ src/pages/view/FilterBar/FilterBar.js
@@ -7,7 +7,7 @@ import {
     acClearItemFilters,
 } from '../../../actions/itemFilters.js'
 import ConfirmActionDialog from '../../../components/ConfirmActionDialog.js'
-import { useOnlineStatus } from '../../../modules/useFakeOnlineStatus.js'
+import { useOnlineStatus } from '../../../online-status-fakes/useFakeOnlineStatus.js'
 import { sGetNamedItemFilters } from '../../../reducers/itemFilters.js'
 import FilterBadge from './FilterBadge.js'
 import classes from './styles/FilterBar.module.css'
