+++ src/pages/view/TitleBar/ActionsBar.js
@@ -17,11 +17,11 @@ import { acSetShowDescription } from '../../../actions/showDescription.js'
 import { apiPostShowDescription } from '../../../api/description.js'
 import ConfirmActionDialog from '../../../components/ConfirmActionDialog.js'
 import DropdownButton from '../../../components/DropdownButton/DropdownButton.js'
-import { OfflineTooltip } from '../../../components/FakeOfflineTooltip.js'
 import MenuItem from '../../../components/MenuItemWithTooltip.js'
 import { useCacheableSection } from '../../../modules/useCacheableSection.js'
-import { useOnlineStatus } from '../../../modules/useFakeOnlineStatus.js'
 import { orObject } from '../../../modules/util.js'
+import { OfflineTooltip } from '../../../online-status-fakes/FakeOfflineTooltip.js'
+import { useOnlineStatus } from '../../../online-status-fakes/useFakeOnlineStatus.js'
 import { sGetDashboardStarred } from '../../../reducers/dashboards.js'
 import { sGetNamedItemFilters } from '../../../reducers/itemFilters.js'
 import { sGetSelected } from '../../../reducers/selected.js'
