+++ src/pages/view/ViewDashboard.js
@@ -19,7 +19,7 @@ import LoadingMask from '../../components/LoadingMask.js'
 import Notice from '../../components/Notice.js'
 import { setHeaderbarVisible } from '../../modules/setHeaderbarVisible.js'
 import { useCacheableSection } from '../../modules/useCacheableSection.js'
-import { useOnlineStatus } from '../../modules/useFakeOnlineStatus.js'
+import { useOnlineStatus } from '../../online-status-fakes/useFakeOnlineStatus.js'
 import { sGetDashboardById } from '../../reducers/dashboards.js'
 import { sGetPassiveViewRegistered } from '../../reducers/passiveViewRegistered.js'
 import { sGetSelectedId } from '../../reducers/selected.js'
